backend/requirements.txt
# FastAPI Framework
fastapi==0.117.1
uvicorn[standard]==0.37.0
starlette==0.48.0

# File Upload Support
python-multipart==0.0.20

# Database Support - Azure SQL Server
SQLAlchemy==2.0.43
pyodbc==5.2.0
# Note: psycopg2-binary removed - not needed for SQL Server

# Security & Cryptography
cryptography==46.0.1
pyOpenSSL==25.3.0

# Data Validation
pydantic==2.11.9
pydantic_core==2.33.2

# Environment & Configuration
python-dotenv==1.0.1

# AI Integration
google-generativeai==0.8.3

# Core Dependencies
annotated-types==0.7.0
anyio==4.11.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
greenlet==3.2.4
h11==0.16.0
idna==3.10
pycparser==2.23
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.15.0

# Railway-specific additions
gunicorn==21.2.0
# Uses existing Azure SQL Server database
